steps: split containers_kube.py into docker.py and kubernetes.py
To make things more maintainable, let's split those step implementations
into distinct files.
