+++ steps/docker.py
@@ -1,4 +1,4 @@
-'''test methods for containers and stuff'''
+'''test methods for docker'''
 
 import re
 from behave import *
@@ -53,24 +53,6 @@ def step_impl(context, rpm, image, host):
                                host,
                                module_args='docker run %s yum install -y %s' % (image, rpm))
 
-@given('"{number}" "{component}" are running')
-def step_impl(context, number, component):
-    '''check expected number of components are running
-       where component is pod, service, node, etc'''
-    import json
-    r = context.remote_cmd('command',
-                            module_args='kubectl get %s -o json' % component)
-    if r:
-        for i in r:
-            data = json.loads(i['stdout'])
-            # terrible hack. must be a better way
-            if 'items' in data:
-                assert len(data['items']) is int(number)
-            else:
-                assert int(number) is 0
-    else:
-        assert False
-
 @when('docker stop container')
 def step_impl(context):
     '''docker stop container'''
@@ -90,27 +72,6 @@ def step_impl(context):
     '''check whether container is running'''
     container_id = get_running_container_id(context)
     assert container_id, "There is not a running container"
-
-@given('nodes are ready')
-def step_impl(context):
-    '''check whether the nodes are ready'''
-    import json
-    r = context.remote_cmd('command', module_args='kubectl get nodes -o json')
-
-    if r:
-        for i in r:
-            data = json.loads(i['stdout'])
-            if 'items' not in data:
-                assert True # trivially true
-            else:
-                # We need to find the Ready condition type and check its status.
-                # See docs/admin/node.md#node-condition in kubernetes repo.
-                for node in data['items']:
-                    for condition in node['status']['conditions']:
-                        if condition['type'] == "Ready":
-                            assert condition['status'] == 'True'
-    else:
-        assert False
 
 @when('docker build an image from "{dockerfile}"')
 @when('docker build an image with tag "{tag}" from "{dockerfile}"')
